pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tot-agent"
version = "0.2.0"
description = "Autonomous browser agent for scripted GUI testing, initially targeting the This-or-That A/B testing platform"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Matt Briggs" }]
keywords = ["browser-agent", "playwright", "anthropic", "a/b-testing", "gui-testing", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.44.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
    "pymdown-extensions>=10.0.0",
]

[project.scripts]
tot-agent = "tot_agent.cli:cli"

[project.urls]
Homepage = "https://github.com/mattbriggs/this-or-that-agent"
Documentation = "https://mattbriggs.github.io/this-or-that-agent"
Repository = "https://github.com/mattbriggs/this-or-that-agent"
Issues = "https://github.com/mattbriggs/this-or-that-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/tot_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--cov=src/tot_agent",
    "--cov-report=html:reports/coverage",
    "--cov-report=term-missing",
    "--cov-report=json:reports/coverage.json",
    "-v",
]

[tool.coverage.run]
source = ["src/tot_agent"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "RUF"]
ignore = ["E501"]
